requirements.txt
Django == 3.0.5
dj-database-url == 0.5.0
psycopg2 
django-polymorphic==2.1.2
django-rest-polymorphic==0.1.9
uwsgi==2.0.18
django-dynamic-raw-id==2.6
djangorestframework==3.11.0
django-filter==2.2.0